requirements-full.txt
albumentations==1.4.24
fastapi==0.115.5
matplotlib==3.10.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
prometheus-fastapi-instrumentator==7.0.0
pydantic==2.10.2
python-Levenshtein==0.26.1
python-multipart==0.0.17
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
streamlit==1.40.1
tqdm==4.67.1
torch==2.5.1
torchvision==0.20.1
uvicorn==0.32.1
requests==2.32.3
